+++ packages/common-widgets/utils/type.ts
@@ -1,13 +1,31 @@
+/* eslint-disable no-redeclare */
 /**
  * @type utils
  * @name types
  */
 
-export function componentWrapper<C, E extends {}>(Component: C, extra: E) {
-    const Comp = Component as C & E;
-    Object.keys(extra).forEach(key => {
-        Comp[key] = extra[key];
-    });
+export function componentWrapper<C, E extends {}>(
+    Component: C,
+    displayName: string,
+    extra?: E,
+): C & E & { displayName?: string };
+export function componentWrapper<C, E extends {}>(Component: C, extra: E): C & E;
+export function componentWrapper<C, E extends {}>(Component, params: string | E, extra?: E) {
+    const Comp = Component as C & E & { displayName?: string };
+
+    if (typeof params === 'string') {
+        Comp.displayName = params;
+        extra &&
+            Object.keys(extra).length &&
+            Object.keys(extra).forEach(key => {
+                Comp[key] = extra[key];
+            });
+    } else {
+        Object.keys(params).forEach(key => {
+            Comp[key] = params[key];
+        });
+    }
+
     return Comp;
 }